Move `live_kit` to `live_kit_client` and add `live_kit_server`
