+++ crates/capture/Cargo.toml
@@ -6,7 +6,8 @@ description = "An example of screen capture"
 
 [dependencies]
 gpui = { path = "../gpui" }
-live_kit = { path = "../live_kit" }
+live_kit_client = { path = "../live_kit_client" }
+live_kit_server = { path = "../live_kit_server" }
 media = { path = "../media" }
 
 anyhow = "1.0.38"
@@ -18,14 +19,10 @@ core-foundation = "0.9.3"
 core-graphics = "0.22.3"
 foreign-types = "0.3"
 futures = "0.3"
-hmac = "0.12"
-jwt = "0.16"
 log = { version = "0.4.16", features = ["kv_unstable_serde"] }
 objc = "0.2"
 parking_lot = "0.11.1"
 postage = { version = "0.4.1", features = ["futures-traits"] }
-serde = { version = "1.0", features = ["derive", "rc"] }
-sha2 = "0.10"
 simplelog = "0.9"
 
 [build-dependencies]
