+++ crates/live_kit_client/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "live_kit"
+name = "live_kit_client"
 version = "0.1.0"
 edition = "2021"
 description = "Bindings to LiveKit Swift client SDK"
 
 [lib]
-path = "src/live_kit.rs"
+path = "src/live_kit_client.rs"
 doctest = false
 
 [dependencies]
